server/requirements.txt
django-admin
djangorestframework
markdown
django-filter
python-dotenv
dj_database_url
psycopg2-binary
requests
# google-cloud-texttospeech